+++ R/write_output.R
@@ -239,7 +239,7 @@ melt_access_spec <- function(dscr, name = NULL){
 #' @noRd
 LagQCKube <- function(allvis, allvistabs, kube){
   qcvals <- getOption("khfunctions.qcvals")
-  stataprikkvals <- grep("^pvern$|^serieprikket$|^naboprikketIomg*", names(kube), value = T)
+  stataprikkvals <- grep("^pvern$|^serieprikket$|^naboprikket$", names(kube), value = T)
   uprikk <- data.table::copy(kube)[, .SD, .SDcols = c(allvistabs, qcvals, stataprikkvals)]
   data.table::setnames(uprikk, qcvals, paste0(qcvals, "_uprikk"))